requirements.txt
PyYAML~=6.0
netmiko~=4.1.2
pysnmp~=4.4.12